datacleaner/API/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
Pillow>=10.0.0
openpyxl==3.1.2

